quran tab in progress till sura of quran.
