+++ lib/home/Quran/sura_details.dart
@@ -1,0 +1,1 @@
+
