+++ linux/flutter/generated_plugins.cmake
@@ -3,7 +3,7 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-        )
+)
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
         )
